Refactor configuration and CLI
